feat: 검증 된 CardDeck 모델 생성
