+++ README.md
@@ -17,8 +17,8 @@
   - [x] 카드는 카드 문양과 숫자의 조합을 반환한다.
 - 카드 더미
   - 카드 더미는 총 52장의 카드로 구성되어 있다.
-    - [ ] ACE 부터 KING까지 13개의 숫자와 4가지 문양을 한 더미로 가지고 있다.
-    - [ ] 플레이어와 딜러는 카드 더미에서 카드를 드로우 할 수 있다.
+    - [x] ACE 부터 KING까지 13개의 숫자와 4가지 문양을 한 더미로 가지고 있다.
+    - [x] 플레이어와 딜러는 카드 더미에서 카드를 드로우 할 수 있다.
 ---
 - 손패
   - 게임의 참여자는 카드 더미에서 뽑힌 카드를 손패로 가지고 있다.
